requirements.txt
appnope==0.1.0
backcall==0.1.0
blessings==1.7
boto3==1.9.93
botocore==1.12.253
Click==7.0
coverage==4.5.2
decorator==4.4.1
dj-database-url==0.5.0
Django==2.1.11
django-configurations==2.1
django-filter==2.1.0
django-model-utils==3.1.2
django-nose==1.4.6
django-storages==1.7.1
django-unique-upload==0.2.1
djangorestframework==3.9.1
docutils==0.15.2
drf-generators==0.4.0
entrypoints==0.3
factory-boy==2.11.1
Faker==2.0.3
flake8==3.7.5
gunicorn==19.9.0
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
livereload==2.6.1
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.0.4
mock==2.0.0
newrelic==4.12.0.113
nose==1.3.7
nose-progressive==1.5.2
parso==0.5.1
pbr==5.4.3
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2-binary==2.7.7
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
python-dateutil==2.8.1
pytz==2018.9
PyYAML==5.1.2
s3transfer==0.2.1
six==1.13.0
text-unidecode==1.3
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
